// src/presentation/pages/overseeing-operations/data-access/waiting-queue/interfaces/waiting-queue-filter-form.interface.ts
import { FormControl } from '@angular/forms';

export interface WaitingQueueFilterFormInterface {
    operation: FormControl<string>;
    date_debut: FormControl<string>;
    date_fin: FormControl<string>;
    numero_demande: FormControl<string>;
    statut: FormControl<string>;
    traitement: FormControl<string>;
    nom_tenant: FormControl<string>;
    initie_par: FormControl<string>;
}